+++ src/features/custom-currency/presets.js
@@ -1,0 +1,61 @@
+export const CURRENCY_PRESETS = {
+    krynn: {
+        currencies: {
+            pp: {
+                label: 'SOSLY.CURRENCY.Krynn.Platinum',
+                abbreviation: 'SOSLY.CURRENCY.Krynn.PlatinumAbbr',
+                icon: 'modules/sosly-5e-house-rules/public/currency/krynn/platinum.webp'
+            },
+            gp: {
+                label: 'SOSLY.CURRENCY.Krynn.Steel',
+                abbreviation: 'SOSLY.CURRENCY.Krynn.SteelAbbr',
+                icon: 'modules/sosly-5e-house-rules/public/currency/krynn/steel.webp'
+            },
+            ep: {
+                label: 'SOSLY.CURRENCY.Krynn.Iron',
+                abbreviation: 'SOSLY.CURRENCY.Krynn.IronAbbr',
+                icon: 'modules/sosly-5e-house-rules/public/currency/krynn/iron.webp'
+            },
+            sp: {
+                label: 'SOSLY.CURRENCY.Krynn.Bronze',
+                abbreviation: 'SOSLY.CURRENCY.Krynn.BronzeAbbr',
+                icon: 'modules/sosly-5e-house-rules/public/currency/krynn/bronze.webp'
+            },
+            cp: {
+                label: 'SOSLY.CURRENCY.Krynn.Silver',
+                abbreviation: 'SOSLY.CURRENCY.Krynn.SilverAbbr',
+                icon: 'modules/sosly-5e-house-rules/public/currency/krynn/silver.webp'
+            }
+        }
+    },
+
+    cormyr: {
+        currencies: {
+            pp: {
+                label: 'SOSLY.CURRENCY.Cormyr.Tricrown',
+                abbreviation: 'SOSLY.CURRENCY.Cormyr.TricrownAbbr',
+                icon: 'systems/dnd5e/icons/currency/platinum.webp'
+            },
+            gp: {
+                label: 'SOSLY.CURRENCY.Cormyr.GoldenLion',
+                abbreviation: 'SOSLY.CURRENCY.Cormyr.GoldenLionAbbr',
+                icon: 'systems/dnd5e/icons/currency/gold.webp'
+            },
+            ep: {
+                label: 'SOSLY.CURRENCY.Cormyr.BlueEye',
+                abbreviation: 'SOSLY.CURRENCY.Cormyr.BlueEyeAbbr',
+                icon: 'systems/dnd5e/icons/currency/electrum.webp'
+            },
+            sp: {
+                label: 'SOSLY.CURRENCY.Cormyr.SilverFalcon',
+                abbreviation: 'SOSLY.CURRENCY.Cormyr.SilverFalconAbbr',
+                icon: 'systems/dnd5e/icons/currency/silver.webp'
+            },
+            cp: {
+                label: 'SOSLY.CURRENCY.Cormyr.CopperThumb',
+                abbreviation: 'SOSLY.CURRENCY.Cormyr.CopperThumbAbbr',
+                icon: 'systems/dnd5e/icons/currency/copper.webp'
+            }
+        }
+    }
+};